chore: rename crac to crc
