+++ Makefile
@@ -1,7 +1,7 @@
 .PHONY: schemas tests style_checks pre_commit_checks
 
 schemas:
-	poetry run datamodel-codegen --input src/renku_crac/api.spec.yaml --input-file-type openapi --output src/schemas/apispec.py --use-double-quotes --target-python-version 3.11 --collapse-root-models --field-constraints --base-class schemas.base.BaseAPISpec
+	poetry run datamodel-codegen --input src/renku_crc/api.spec.yaml --input-file-type openapi --output src/schemas/apispec.py --use-double-quotes --target-python-version 3.11 --collapse-root-models --field-constraints --base-class schemas.base.BaseAPISpec
 
 style_checks:
 	poetry check
@@ -14,7 +14,7 @@ style_checks:
 tests:
 	@rm -f .tmp.pid coverage.lcov .coverage data_services.db
 	-poetry run pytest
-	DUMMY_STORES=true poetry run coverage run -a -m sanic --debug --single-process renku_crac.main:create_app --factory & echo $$! > .tmp.pid
+	DUMMY_STORES=true poetry run coverage run -a -m sanic --debug --single-process renku_crc.main:create_app --factory & echo $$! > .tmp.pid
 	@sleep 10
 	-poetry run st run http://localhost:8000/api/data/spec.json --validate-schema True --checks all --hypothesis-max-examples 20 --data-generation-method all --show-errors-tracebacks --hypothesis-suppress-health-check data_too_large --max-response-time 100 -v --header "Authorization: bearer some-random-key-123456"
 	cat .tmp.pid | xargs kill
