+++ README.md
@@ -8,7 +8,7 @@ A set of services that handle reading and writing data from Postgres about compu
 
 1. `poetry install`
 2. `pre-commit install` to install pre commit hooks
-3. `DUMMY_STORES=true poetry run python src/renku_crac/main.py --debug --dev --fast`
+3. `DUMMY_STORES=true poetry run python src/renku_crc/main.py --debug --dev --fast`
 
 ## Developing
