+++ src/models/__init__.py
@@ -1,3 +1,3 @@
 """Domain models and exceptions."""
 # flake8: noqa
-from models.crac import APIUser, Authenticator, GpuKind, Quota, ResourceClass, ResourcePool, User, UserStore
+from models.crc import APIUser, Authenticator, GpuKind, Quota, ResourceClass, ResourcePool, User, UserStore
