--- tests/unit/renku_crac/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for renku_crac package."""
