+++ tests/unit/renku_crc/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for renku_crc package."""
